Moved workflows to Models/Workflows and added helper method `runWorkflow` to invoke a workflow.
