+++ lib/Models/Workflows/SelectableDimensionWorkflow.ts
@@ -4,7 +4,10 @@ import { BaseModel } from "../Definition/Model";
 import {
   SelectableDimensionGroup,
   SelectableDimension
-} from "./SelectableDimensions";
+} from "../SelectableDimensions/SelectableDimensions";
+import ViewState from "../../ReactViewModels/ViewState";
+import Terria from "../Terria";
+import { runInAction } from "mobx";
 
 /**
  * Model for SelectableDimensionWorkflow. Basically just includes a bunch of selectableDimensions, name, icon and a catalog member.
@@ -12,11 +15,14 @@ import {
  *
  * See `lib/ReactViews/Workflow/Workflows/SelectableDimension/SelectableDimensionWorkflow.tsx` for more info and usage
  */
-export default interface SelectableDimensionWorkflow {
+interface SelectableDimensionWorkflow {
   /** Human readable name - used as title */
   name: string;
   icon: IconProps["glyph"];
+
+  /** Item to which this workflow belongs **/
   item: BaseModel;
+
   onClose?: () => void;
   /** Footer button */
   footer?: { onClick: () => void; buttonText: string };
@@ -24,6 +30,29 @@ export default interface SelectableDimensionWorkflow {
   /** This allows up to two levels of SelectableDimensionGroup */
   selectableDimensions: SelectableDimensionWorkflowGroup[];
 }
+
+namespace SelectableDimensionWorkflow {
+  /**
+   * Runs a selectable dimension workflow which is a workflow for a workbench item.
+   *
+   * @param viewStateOrTerria - The {@link ViewState} or {@link Terria} instance
+   * @param workflow - A {@link SelectableDimensionWorkflow} instance
+   */
+  export function runWorkflow(
+    viewStateOrTerria: ViewState | Terria,
+    workflow: SelectableDimensionWorkflow
+  ) {
+    runInAction(() => {
+      const terria =
+        viewStateOrTerria instanceof Terria
+          ? viewStateOrTerria
+          : viewStateOrTerria.terria;
+      terria.selectableDimensionWorkflow = workflow;
+    });
+  }
+}
+
+export default SelectableDimensionWorkflow;
 
 /** This is essentially the same as `SelectableDimensionGroup`, but allows two levels of nested `SelectableDimensionGroup`, instead of one */
 export interface SelectableDimensionWorkflowGroup
