+++ lib/Models/Workflows/VectorStylingWorkflow.ts
